Move checkers into VersionUpdate\Checker namespace
