+++ src/VersionUpdater/Checker/IntervalChecker.php
@@ -9,7 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Psy\VersionUpdater;
+namespace Psy\VersionUpdater\Checker;
+
+use Psy\VersionUpdater\Checker;
 
 class IntervalChecker extends GitHubChecker
 {
